separate player database
