+++ ST_database_definition.py
@@ -1,0 +1,30 @@
+from collections import defaultdict
+from loguru import logger
+
+# begin definition of default dict keys for players
+CALLERNAME = "Caller_name"
+FIRSTCALL = "First_call"
+RECENTCALL = "Recent_Call"
+PLUS_ONES = "Plus_one"
+PARTNER_HISTORY = "Partners_history"
+MESSAGE_HISTORY = "Message_history"
+CURRENT_TABLE_ASSIGNMENT = "Current_Table"
+CURRENT_TEAM_NAME = "Current_team"
+
+@logger.catch
+def dict_default():
+    sample_player_dict = {
+        CALLERNAME: "",
+        CURRENT_TABLE_ASSIGNMENT: "Undefined",
+        CURRENT_TEAM_NAME: "notset",
+        PLUS_ONES: int(0),  # represents number of extra seats reserved at the table
+        PARTNER_HISTORY: [],
+        MESSAGE_HISTORY: [],
+        FIRSTCALL: None,
+        RECENTCALL: None,
+    }
+    return sample_player_dict
+
+
+players_database = defaultdict(dict_default)
+players_database["root"] = "root"
